requirements.txt
python-ldap
requests
glog
fire
httpretty
sure
nose
